refactor(wizard-tools): split into shared core + MCP facade
wizard-tools.ts becomes a folder: tools.ts holds the shared behavior
(skills, env helpers, ask-cap policy, vault plumbing, audit ledger,
WIZARD_TOOL_NAMES), mcp.ts is the MCP facade (zod declarations +
createWizardToolsServer), index.ts re-exports both so the existing
@lib/wizard-tools import path is unchanged. The pi facade now imports
from @lib/wizard-tools/tools — both harness facades sit over one core,
and neither depends on the other's protocol layer. No behavior change.

Generated-By: PostHog Code
Task-Id: c8dcf8d0-3565-4c83-8929-3c7faecdcb1c
